requirements.txt
requests
beautifulsoup4
nltk
mysql-connector-python
gcsa